StudyMate-Flask/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymupdf==1.24.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
ibm-watsonx-ai==1.1.16
tqdm==4.66.4
